requirements.txt
pywin32==306
pillow==10.2.0
pyinstaller==6.3.0 